# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Assessment",
        "description": "Evermos assessment test",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "adhityanugraha.jtk@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/soldier/initMagazine": {
            "post": {
                "tags": [
                    "soldier"
                ],
                "summary": "Magazine Initialization",
                "description": "Initialize magazazine that the soldier carry to the war.",
                "operationId": "App\\Http\\Controllers\\Soldier\\MagazineController::initMagazine",
                "requestBody": {
                    "description": "initialize value for the magazine",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "magazineNumber": {
                                        "description": "Soldier carry the magazine into the war",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "magazineSize": {
                                        "description": "Soldier carry the magazine into the war",
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
}